live-api-minimal-test/requirements.txt
# Minimal deps for testing Gemini Live API (no FastAPI, no browser).
# Install: pip install -r requirements.txt
google-genai>=1.65.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
